first attempt after POC of creating a scopeSlot base cards component to enable us to provide multiple optionionated wrapper components / layouts for the multiple uses of f-content-cards
