+++ packages/f-content-cards/src/components/layouts/_tests/PostOrderLayout.test.js
@@ -1,0 +1,18 @@
+describe('PostOrderLayout', () => {
+    allure.feature('Post Order Layout');
+    // it('should emit a custom content callback when mounted', () => {
+    //     // Arrange & Act
+    //     const instance = shallowMount(ContentCards, {
+    //         propsData: {
+    //             apiKey,
+    //             userId
+    //         }
+    //     });
+    //
+    //     // Assert
+    //     expect(instance.emitted()['custom-cards-callback']).toBeTruthy(); // Correct event emitted
+    //     expect(instance.emitted()['custom-cards-callback'].length).toBe(1); // Once
+    //     expect(instance.emitted()['custom-cards-callback'][0].length).toBe(1); // With one argument
+    //     expect(instance.emitted()['custom-cards-callback'][0][0]).toBeInstanceOf(Function); // That is a callback
+    // });
+});
